Optimized painting to eliminate flickering apple
Because we were loading the gif of the apple every time we painted the
canvas, it was causing a flickering. The loading of the gif has been
moved into the startup method and only the placement of the image on
the canvas happens when the canvas is re-painted. The flickering no
longer occurs.